refactor: promote deep_qa workflow from contributed/ to factory/workflow/
deep_qa is a first-class workflow, not a contributed one. Move it alongside
definitions.py and update imports in definitions.py and test_workflow_qa.py.
